+++ lib/src/extensions/ble_device_extension.dart
@@ -1,4 +1,3 @@
-import 'package:universal_ble/src/universal_ble_pigeon/universal_ble.g.dart';
 import 'package:universal_ble/src/utils/cache_handler.dart';
 import 'package:universal_ble/universal_ble.dart';
 
@@ -19,7 +18,8 @@ extension BleDeviceExtension on BleDevice {
   /// Connects to the device.
   /// [autoConnect] enables automatic reconnection when the device becomes available.
   Future<void> connect({bool autoConnect = false, Duration? timeout}) =>
-      UniversalBle.connect(deviceId, autoConnect: autoConnect, timeout: timeout);
+      UniversalBle.connect(deviceId,
+          autoConnect: autoConnect, timeout: timeout);
 
   /// Disconnects from the device.
   Future<void> disconnect() => UniversalBle.disconnect(deviceId);