+++ lib/src/extensions/ble_service_extension.dart
@@ -1,4 +1,3 @@
-import 'package:universal_ble/src/universal_ble_pigeon/universal_ble.g.dart';
 import 'package:universal_ble/universal_ble.dart';
 
 /// Extension methods for [BleService] objects.